Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pyjwt==2.9.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
